Refactor 1.0
