--- EventService/src/main/java/com/enfec/sb/eventapi/WebConfig.java
@@ -1,38 +1,0 @@
-package com.enfec.sb.eventapi;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@EnableWebMvc
-public class WebConfig implements WebMvcConfigurer {
-    /**
-     * {@inheritDoc}
-     */
-	@Override
-	public void addCorsMappings(CorsRegistry registry) {
-
-		registry.addMapping("/**").allowedOrigins("*").allowedMethods("GET", "POST", "OPTIONS", "PUT")
-				.allowedHeaders("Content-Type", "X-Requested-With", "accept", "Origin", "Access-Control-Request-Method",
-						"Access-Control-Request-Headers")
-				.exposedHeaders("Access-Control-Allow-Origin", "Access-Control-Allow-Credentials")
-				.allowCredentials(true).maxAge(3600);
-		
-	}
-	
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/static/**").addResourceLocations("classpath:/static/");
-        registry.addResourceHandler("swagger-ui.html")
-                .addResourceLocations("classpath:/META-INF/resources/");
-
-        registry.addResourceHandler("/webjars/**")
-                .addResourceLocations("classpath:/META-INF/resources/webjars/");    
-   }
-}
