+++ EventService/src/test/java/com/enfec/eventapi/EventApiApplicationTests.java
@@ -1,0 +1,14 @@
+package com.enfec.eventapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EventApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+	
+
+}
