--- EventService/src/test/java/com/enfec/sb/eventapi/EventApiApplicationTests.java
@@ -1,14 +1,0 @@
-package com.enfec.sb.eventapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EventApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-	
-
-}
