--- refundService/src/main/java/com/enfec/sb/refundapi/RefundApiApplication.java
@@ -1,41 +1,0 @@
-package com.enfec.sb.refundapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.context.annotation.Bean;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-/**
- * API Application initializer for refund
- */
-@SpringBootApplication
-@EnableSwagger2
-public class RefundApiApplication extends SpringBootServletInitializer {
-
-    /**
-     * Main function
-     */
-    public static void main(String[] args) {
-        SpringApplication.run(RefundApiApplication.class, args);
-
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected SpringApplicationBuilder configure(SpringApplicationBuilder builder) {
-        return builder.sources(RefundApiApplication.class);
-    }
-
-    @Bean
-    public Docket productApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select().apis(RequestHandlerSelectors.basePackage("com.enfec")).build();   
-    }
-}
