+++ refundService/src/test/java/com/enfec/refundapi/RefundApiApplicationTests.java
@@ -1,0 +1,14 @@
+package com.enfec.refundapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RefundApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+	
+
+}
