--- refundService/src/test/java/com/enfec/sb/refundapi/RefundApiApplicationTests.java
@@ -1,14 +1,0 @@
-package com.enfec.sb.refundapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RefundApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-	
-
-}
